requirements.txt
# great_expectations[sqlalchemy]==0.17.11
pytest==7.4.0
faker==19.3.0
delta-spark==2.3.0
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
python-dotenv==0.21.0
mypy==0.991
flake8==6.0.0
black==22.12.0
isort==5.12.0
soda-core-spark-df==3.3.1
prefect==2.20
pydantic==2.9.2
pydantic_core==2.23.4
boto3